requirements.txt
absl-py==0.6.1
astor==0.7.1
bleach==1.5.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.0.2
Flask-Cors==3.0.7
gast==0.2.0
grpcio==1.17.1
h5py==2.9.0
html5lib==0.9999999
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
numpy==1.15.4
opencv-python==3.4.4.19
Pillow==5.3.0
protobuf==3.6.1
pyparsing==2.3.0
python-dateutil==2.7.5
requests==2.21.0
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
